+++ QuizRush.Web/Program.cs
@@ -1,0 +1,32 @@
+using QuizRush.Web.Components;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+builder.Services.AddRazorComponents()
+    .AddInteractiveServerComponents();
+builder.Services.AddScoped(_ => new HttpClient
+{
+    BaseAddress = new Uri("http://localhost:5176")
+});
+builder.Services.AddScoped<QuizRush.Web.Services.GameService>();
+builder.Services.AddScoped<QuizRush.Web.Services.AuthService>();
+builder.Services.AddScoped<QuizRush.Web.Services.LocalStorageService>();
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (!app.Environment.IsDevelopment())
+{
+    app.UseExceptionHandler("/Error", createScopeForErrors: true);
+    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+    app.UseHsts();
+}
+
+app.UseAntiforgery();
+
+app.MapStaticAssets();
+app.MapRazorComponents<App>()
+    .AddInteractiveServerRenderMode();
+
+app.Run();